+++ output/changelog/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- Generated at: 2025-12-06T22:34:05.816889
-- Total time: 31.59 seconds
+- Generated at: 2025-12-07T11:54:13.178431
+- Total time: 215.40 seconds
 - Features: 7
 - Models: 20